rename/style: common package 변경, 메서드명 변경
