+++ src/main/java/io/hhplus/server_construction/domain/payment/exception/PaymentException.java
@@ -1,6 +1,6 @@
 package io.hhplus.server_construction.domain.payment.exception;
 
-import io.hhplus.server_construction.common.exception.BaseException;
+import io.hhplus.server_construction.support.exception.BaseException;
 
 public class PaymentException extends BaseException {
