+++ src/main/java/io/hhplus/server_construction/domain/waiting/exceprtion/WaitingException.java
@@ -1,6 +1,6 @@
 package io.hhplus.server_construction.domain.waiting.exceprtion;
 
-import io.hhplus.server_construction.common.exception.BaseException;
+import io.hhplus.server_construction.support.exception.BaseException;
 
 public class WaitingException extends BaseException {
