+++ src/main/java/io/hhplus/server_construction/support/config/SwaggerConfig.java
@@ -1,14 +1,10 @@
-package io.hhplus.server_construction.common.config;
+package io.hhplus.server_construction.support.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.servers.Server;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import java.util.List;
 
 @Configuration
 @RequiredArgsConstructor
