+++ src/main/java/io/hhplus/server_construction/support/log/ResponseInfoLogData.java
@@ -1,4 +1,4 @@
-package io.hhplus.server_construction.common.log;
+package io.hhplus.server_construction.support.log;
 
 import jakarta.servlet.http.HttpServletResponse;
