+++ src/test/java/io/hhplus/server_construction/application/reservation/facade/ReservationFacadeTest.java
@@ -12,14 +12,10 @@ import io.hhplus.server_construction.domain.concert.vo.ConcertScheduleStatus;
 import io.hhplus.server_construction.domain.concert.vo.ConcertSeatGrade;
 import io.hhplus.server_construction.domain.concert.vo.ConcertSeatStatus;
 import io.hhplus.server_construction.domain.reservation.Reservation;
-import io.hhplus.server_construction.domain.reservation.ReservationItem;
 import io.hhplus.server_construction.domain.reservation.service.ReservationService;
 import io.hhplus.server_construction.domain.reservation.vo.ReservationStatus;
 import io.hhplus.server_construction.domain.user.User;
 import io.hhplus.server_construction.domain.user.service.UserService;
-import io.hhplus.server_construction.domain.waiting.exceprtion.WaitingException;
-import io.hhplus.server_construction.domain.waiting.exceprtion.WaitingExceptionEnums;
-import io.hhplus.server_construction.domain.waiting.service.WaitingService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -56,7 +52,7 @@ class ReservationFacadeTest {
 
     @Test
     @DisplayName("유효하지 않은 토큰으로 콘서트 좌석을 예약한다.")
-    void reservationConcertTokenException() {
+    void setConcertReservationTokenException() {
         // given
         Long userId = 1L;
         ReservationConcertCommand reservationConcertCommand = new ReservationConcertCommand(List.of(1L, 2L, 3L), userId);
@@ -88,8 +84,8 @@ class ReservationFacadeTest {
 
         // when
         when(userService.findUserById(reservationConcertCommand.userId())).thenReturn(user);
-        when(concertService.reservationSeat(reservationConcertCommand.concertSeatIdList())).thenReturn(concertSeatList);
-        when(reservationService.reservationConcert(concertSeatList, user)).thenReturn(Reservation.create(
+        when(concertService.setSeatReservation(reservationConcertCommand.concertSeatIdList())).thenReturn(concertSeatList);
+        when(reservationService.setConcertReservation(concertSeatList, user)).thenReturn(Reservation.create(
                 1L,
                 user,
                 BigDecimal.valueOf(3000),
@@ -97,7 +93,7 @@ class ReservationFacadeTest {
                 anyList(),
                 now
         ));
-        ReservationConcertResult result = reservationFacade.reservationConcert(reservationConcertCommand);
+        ReservationConcertResult result = reservationFacade.setConcertReservation(reservationConcertCommand);
 
         // then
         assertThat(result).isNotNull();
@@ -105,7 +101,7 @@ class ReservationFacadeTest {
 
     @Test
     @DisplayName("유효한 상태가 아닌 콘서트를 예약한다.")
-    void reservationConcertNotPossible() {
+    void setConcertReservationNotPossible() {
         // given
         Long userId = 1L;
         List<Long> concertSeatIdList = List.of(1L, 2L, 3L);
@@ -113,10 +109,10 @@ class ReservationFacadeTest {
 
         // when
         when(userService.findUserById(userId)).thenReturn(any());
-        when(concertService.reservationSeat(concertSeatIdList)).thenThrow(new ConcertException(ConcertExceptionEnums.ALREADY_RESERVATION));
+        when(concertService.setSeatReservation(concertSeatIdList)).thenThrow(new ConcertException(ConcertExceptionEnums.ALREADY_RESERVATION));
 
         // then
-        assertThatThrownBy(() -> reservationFacade.reservationConcert(reservationConcertCommand))
+        assertThatThrownBy(() -> reservationFacade.setConcertReservation(reservationConcertCommand))
                 .isInstanceOf(ConcertException.class)
                 .hasMessageContaining(ConcertExceptionEnums.ALREADY_RESERVATION.getMessage());
     }
