+++ src/test/java/io/hhplus/server_construction/domain/reservation/service/ReservationServiceTest.java
@@ -84,7 +84,7 @@ class ReservationServiceTest {
 
     @Test
     @DisplayName("콘서트를 예약한다. - 좌석의 총 합계 계산")
-    void reservationConcert() {
+    void setConcertReservation() {
         // given
         List<ConcertSeat> concertSeatList = List.of(this.concertSeat1, this.concertSeat2, this.concertSeat3);
         BigDecimal totalPrice = concertSeatList.stream()
@@ -95,7 +95,7 @@ class ReservationServiceTest {
         // when
         when(reservationRepository.saveReservation(reservation)).thenReturn(reservation);
         when(reservationRepository.saveAllReservationItems(any())).thenReturn(any());
-        Reservation result = reservationService.reservationConcert(concertSeatList, this.user);
+        Reservation result = reservationService.setConcertReservation(concertSeatList, this.user);
 
         // then
         assertThat(result.getTotalPrice()).isEqualTo(totalPrice);
